requirements.txt
certifi==2019.6.16
chardet==3.0.4
httplib2==0.13.1
idna==2.8
oauth2==1.9.0.post1
oauthlib==3.1.0
PySocks==1.7.0
python-dotenv==0.10.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
tweepy==3.8.0
urllib3==1.25.3
